Miniconda3 is included.     1. The default location is ~/.miniconda3     2. For installing packages to be used by system programs the     installation should be done in ~/.local/lib/python3.x using system     pip/pip3. Please check install.sh in this regard.

+++ install.sh
@@ -25,14 +25,10 @@ do
 	sudo -n true; sleep 60; kill -0 "$$" || exit; 
 done 2>/dev/null &
 
-action "Update the apt and installing curl"
+action "Update the apt and installing download managers"
 sudo -E apt-get update && sudo -E apt-get upgrade -ym
-sudo -E apt-get install -ym curl
+sudo -E apt-get install -ym curl aria2
 ok
-
-# action "Adding repository for latest nodejs and npm"
-# curl -sL https://deb.nodesource.com/setup_8.x | sudo -E bash -
-# ok
 
 action "Installing required packages"
 step "Update the apt "
@@ -44,16 +40,17 @@ sudo -E apt-get install -ym zsh tmux sed xsel stow neovim direnv fonts-noto-core
 # for C/C++ development
 sudo -E apt-get install -ym build-essential clang clang-tools clang-tidy \
 	global universal-ctags cmake ccache git
-# for python 
-sudo -E apt install -ym python3-pip python3-tk 
-# for nodejs
-sudo -E apt install -ym nodejs
+# for python
+sudo -E apt install -ym python3-pip python3-tk
 # for markdown and latex
-sudo -E apt-get install -ym pandoc markdown texlive dvipng \
-	texlive-latex-extra texlive-formats-extra texlive-publishers composer \
-	texlive-bibtex-extra biber texlive-font-utils chktex tidy odt2txt dos2unix
+sudo -E apt-get install -ym pandoc markdown texlive dvipng texlive-luatex \
+	texlive-latex-extra texlive-formats-extra texlive-publishers \
+	texlive-science texworks texlive-bibtex-extra biber texlive-font-utils \
+	chktex tidy odt2txt dos2unix
+# for php composer framework
+sudo -E apt-get install -ym composer
 #Install common truetype font
-sudo -E apt-get install ttf-mscorefonts-installer
+sudo -E apt-get install -ym ttf-mscorefonts-installer
 # for english dictionary
 #sudo -E apt install -ym dictd dict-gcide dict-vera dict-jargon dict-elements \
 #	dict-moby-thesaurus dict
@@ -61,6 +58,7 @@ sudo -E apt-get install ttf-mscorefonts-installer
 sudo -E apt install -ym proj-bin libproj-dev gdal-bin libgdal-dev python3-gdal \
 	libgeos++-dev libgeos-dev
 ok
+
 step "Removing unnecessary packages"
 sudo -E apt-get autoremove -ym
 sudo -E apt-get autoclean -ym
@@ -76,20 +74,20 @@ sudo -E chown -R "$USER:$USER" ~/.local/include
 sudo -E chown -R "$USER:$USER" ~/.local/share
 ok
 
-step "Installing python packages"
-pip3 install --user -U cython
-pip3 install --user -U numpy sympy scipy statsmodels scikit-learn dask \
-	tensorflow pywavelets pandas xarray geopandas pysal pyresample pillow \
-	matplotlib bokeh holoviews seaborn cartopy numba nose netcdf4 \
-	tables h5py xlwt ipython jupyter ipywidgets notebook jedi psutil \
-	setproctitle yamllint proselint demjson scrapy beautifulsoup4 notedown
-pip3 install --user -U orange3 glueviz
-
-# GDAL support - Install this way only if python3-GDAL fail
-#pip3 install --user -U GDAL==$(gdal-config --version) --global-option=build_ext \
-#	--global-option=-I/usr/include/gdal
-# If cartopy cause Ipython to crash: hint: try ax.coastlines()
-#pip3 install --user -U shapely cartopy --no-binary shapely --no-binary cartopy
+step "Installing python packages for system python"
+# Note: This file also setup miniconda. The idea is non-data science package 
+# that is needed by user doftware like neovim is installed outside miniconda,
+# so they are available does not matter which conda environment is active.
+#
+# Further, these package can be also import in miniconda as miniconda python 
+# has ~/.local/lib/python3.x in the path. Beware the libraries installed in 
+# ~/.local/lib/python3.x has higher preference in minconda so do not install 
+# any data science packages in ~/.local/lib/python3.x, otherwise sub-optimal
+# performance in data science workload might be observed.
+SYSPIP=/usr/bin/pip
+SYSPIP3=/usr/bin/pip3
+$SYSPIP3 install --user -U proselint yamllint nose pytest jedi psutil \
+	setproctitle demjson ipython tqdm
 ok
 
 action "Configuring stow"
@@ -212,13 +210,13 @@ ok
 
 step "Enabling python support in nvim"
 #pip install --user -U neovim
-pip install --user -U pynvim
-pip3 install --user -U pynvim
+$SYSPIP install --user -U pynvim
+$SYSPIP3 install --user -U pynvim
 
 # Required by ncm2-ultisnips
-pip3 install --user -U notedown
+$SYSPIP3 install --user -U notedown
 # Required by vimtex
-pip3 install --user -U neovim-remote
+$SYSPIP3 install --user -U neovim-remote
 ok
 
 # step "Enabling ruby support in nvim"
@@ -229,29 +227,65 @@ step "Installing vim plugins"
 nvim +PlugInstall +PlugUpdate +UpdateRemotePlugins +PlugUpgrade +PlugClean +qa
 ok
 
-step "Fixing spell check in vim"
-if [ "$(ls -A ~/.local/share/nvim/plugged/vim-grammarous/misc)" ]; then
-	echo "If you are facing problem with spell check in vim"
-	echo "Delete the old languagetools"
-	rm -iRf ~/.local/share/nvim/plugged/vim-grammarous/misc
-	nvim +GrammarousCheck +qa
-	cd ~/.local/share/nvim/plugged/LanguageClient-neovim/
-	rm -f bin/languageclient
-	bash install.sh
-	cd -
-fi
-ok
-
 step "Installing Language Servers"
 # Language Server for Python
-pip3 install --user -U python-language-server
+$SYSPIP3 install --user -U python-language-server
 # Language server for javascript and typescript
+## install nodejs
+#sudo -E apt install -ym nodejs
 # npm config set prefix $HOME
 # export PATH=$HOME/node_modules/.bin:$PATH
 # npm install -g javascript-typescript-langserver
 # Language server for PHP is installed by vim during plugin install
 ok
+
+step "Fixing spell check in vim if any"
+read -p "Do you face any problem for grammar check in vim?(Y/N) " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	if [ "$(ls -A ~/.local/share/nvim/plugged/vim-grammarous/misc)" ]; then
+		echo "If you are facing problem with spell check in vim"
+		echo "Delete the old languagetools"
+		rm -iRf ~/.local/share/nvim/plugged/vim-grammarous/misc
+		nvim +GrammarousCheck +qa
+		cd ~/.local/share/nvim/plugged/LanguageClient-neovim/
+		rm -f bin/languageclient
+		bash install.sh
+		cd -
+	fi
+else
+	info "No grammar check problem as per user."
+fi
+ok
 info "nvim/vim configuration is complete"
+
+action "Setting up miniconda3 environment"
+read -p "Do you want to setup miniconda3?(Y/N) " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	step "Installing/updating miniconda3"
+	curl https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh --output ~/miniconda3.sh
+	chmod +x ~/miniconda3.sh
+	mkdir -p ~/.miniconda3
+	bash ~/miniconda3.sh -b -u -p ~/.miniconda3
+	rm -f ~/miniconda3.sh
+	ok
+	
+	step "Setting up conda"
+	stow -t ~ -R conda
+	# By now ~/.shell_common_config is present and can be loaded to init conda
+	source ~/.shell_common_config
+	conda activate
+	conda install -y numpy scipy statsmodels pandas xarray geopandas matplotlib \
+		cartopy ipython jupyter ipywidgets notebook nb_conda_kernels ipykernel \
+		jupyter_contrib_nbextensions
+	conda install h5py netcdf4 orange3 glueviz bottleneck
+else
+	info "User opted to not set up miniconda3."
+fi
+ok
 
 step "Applying base16 brewer theme"
 bash -lic base16_brewer
